Reorganized components into namespaces matching Bootstrap's categorization.
