+++ Blazorade.Bootstrap.Components.Showroom/_Imports.razor
@@ -5,3 +5,6 @@
 @using Blazorade.Bootstrap.Components
 @using Blazorade.Bootstrap.Components.Builder
 @using Blazorade.Bootstrap.Components.Model
+@using Blazorade.Bootstrap.Components.Content
+@using Blazorade.Bootstrap.Components.Layout
+@using Blazorade.Bootstrap.Components.Utilities
