+++ Blazorade.Bootstrap.Components/BootstrapComponentBase.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.JSInterop;
+using Blazorade.Bootstrap.Components.Utilities;
 
 namespace Blazorade.Bootstrap.Components
 {
