+++ Blazorade.Bootstrap.Components/Layout/Row.razor.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Blazorade.Bootstrap.Components
+namespace Blazorade.Bootstrap.Components.Layout
 {
     partial class Row
     {
