+++ Blazorade.Bootstrap.Components/Modal.razor
@@ -11,7 +11,7 @@
             {
                 <ModalHeader>
                     <Heading Level="HeadingLevel.H5" class="@ClassNames.Modals.Title">@this.Header</Heading>
-                    <CloseModalButton />
+                    <CloseIcon data-dismiss="modal" />
                 </ModalHeader>
             }
