+++ Blazorade.Bootstrap.Components/Utilities/CloseIcon.razor
@@ -1,4 +1,4 @@
 ﻿
-<button type="button" class="close" data-dismiss="modal">
+<button type="button" class="close" aria-label="Close">
     <span aria-hidden="true">&times;</span>
 </button>
