+++ Blazorade.Bootstrap.Components/_Imports.razor
@@ -1,2 +1,7 @@
 ﻿@using Microsoft.AspNetCore.Components.Web
+@using Blazorade.Bootstrap.Components.Content
+@using Blazorade.Bootstrap.Components.Layout
 @using Blazorade.Bootstrap.Components.Model
+@using Blazorade.Bootstrap.Components.Content
+@using Blazorade.Bootstrap.Components.Layout
+@using Blazorade.Bootstrap.Components.Utilities
