+++ ClientTestHost/_Imports.razor
@@ -7,4 +7,8 @@
 @using ClientTestHost.Shared
 
 @using Blazorade.Bootstrap.Components
+@using Blazorade.Bootstrap.Components.Builder
 @using Blazorade.Bootstrap.Components.Model
+@using Blazorade.Bootstrap.Components.Content
+@using Blazorade.Bootstrap.Components.Layout
+@using Blazorade.Bootstrap.Components.Utilities
